pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Gaevskii.V <gaevskii.v95@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
uvicorn = "^0.31.0"
django-bootstrap5 = "^24.3"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
django-extensions = "^3.2.3"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
